requirements.txt
# General Dependencies 
python-dotenv==1.0.1

# LLM Dependencies
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.22
openai-whisper==20231117
torch
torchaudio

# Record audio
scipy==1.14.1
sounddevice==0.5.0